ATO-47: move ais entities
