--- orchestration-shared/src/main/java/uk/gov/di/orchestration/entity/AccountInterventionResponse.java
@@ -1,6 +1,0 @@
-package uk.gov.di.orchestration.entity;
-
-import com.google.gson.annotations.Expose;
-
-public record AccountInterventionResponse(
-        @Expose AccountInterventionStatus state, @Expose String auditLevel) {}
