--- orchestration-shared/src/main/java/uk/gov/di/orchestration/entity/AccountInterventionStatus.java
@@ -1,10 +1,0 @@
-package uk.gov.di.orchestration.entity;
-
-import com.google.gson.annotations.Expose;
-import com.google.gson.annotations.SerializedName;
-
-public record AccountInterventionStatus(
-        @Expose boolean blocked,
-        @Expose boolean suspended,
-        @Expose @SerializedName("reproveIdentity") boolean reproveIdentity,
-        @Expose @SerializedName("resetPassword") boolean resetPassword) {}
